修改html上传作为单独tab，优化文件结构
