+++ src/pages/user/Sync/ProxySyncSection.tsx
@@ -24,32 +24,45 @@ import { useEffect, useState } from "react";
 import { useIdle, useMediaQuery } from "@mantine/hooks";
 import { IconAlertCircle, IconDownload, IconHelp, IconRepeat } from "@tabler/icons-react";
 import { navigate } from "vike/client/router";
-import { Game } from "@/types/game";
 import { openAlertModal } from "@/utils/modal";
 import { checkProxy } from "@/utils/checkProxy.ts";
 import { getCrawlStatus, getUserCrawlToken } from "@/utils/api/user.ts";
+import { getCrawlStatistic } from "@/utils/api/misc.ts";
+import useGame from "@/hooks/useGame.ts";
 import useShellViewportSize from "@/hooks/useShellViewportSize.ts";
-import { CrawlStatisticProps, CrawlStatusProps, SyncResultSnapshot } from "@/types/sync";
+import { Game } from "@/types/game";
+import { ScoreChangesProps } from "@/pages/user/Sync";
 import { RadioCardGroup } from "@/components/RadioCardGroup.tsx";
 import { CopyButtonWithIcon } from "@/components/Sync/CopyButtonWithIcon.tsx";
 import { CrawlTokenAlert } from "@/components/Sync/CrawlTokenAlert.tsx";
 import { WechatOAuthLink } from "@/components/Sync/WechatOAuthLink.tsx";
+import { ScoresChangesModal } from "@/components/Sync/ScoresChangesModal.tsx";
 import classes from "../Sync.module.css";
 
-interface ProxySyncSectionProps {
-  crawlStatistic: CrawlStatisticProps | null;
-  game: Game;
-  onGameChange: (value: Game) => void;
-  onSyncResult: (result: SyncResultSnapshot) => void;
-  onViewSyncResult: (result: SyncResultSnapshot) => void;
+interface CrawlStatisticProps {
+  success_rate: number;
+  average_crawl_time: number;
 }
 
-export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncResult, onViewSyncResult }: ProxySyncSectionProps) => {
+interface CrawlStatusProps {
+  game: Game;
+  friend_code: number;
+  status: string;
+  create_time: string;
+  complete_time: string;
+  scores: ScoreChangesProps[];
+  failed_difficulties: number[];
+}
+
+export const ProxySyncSection = () => {
   const [isProxyAvailable, setIsProxyAvailable] = useState(false);
   const [hasNetworkError, setHasNetworkError] = useState(false);
   const [crawlToken, setCrawlToken] = useState<string | null>(null);
+  const [crawlStatistic, setCrawlStatistic] = useState<CrawlStatisticProps | null>(null);
   const [crawlStatus, setCrawlStatus] = useState<CrawlStatusProps | null>(null);
+  const [resultModalOpened, setResultModalOpened] = useState(false);
   const [step, setStep] = useState(0);
+  const [game, setGame] = useGame();
   const idle = useIdle(60000);
   const isLoggedOut = !localStorage.getItem("token");
   const { width } = useShellViewportSize();
@@ -75,6 +88,23 @@ export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncRes
       loadCrawlToken();
     }
   }, []);
+
+  useEffect(() => {
+    const getCrawlStatisticHandler = async () => {
+      try {
+        const res = await getCrawlStatistic(game);
+        const data = await res.json();
+        if (!data.success) {
+          throw new Error(data.message);
+        }
+        setCrawlStatistic(data.data);
+      } catch (error) {
+        console.log(error);
+      }
+    };
+
+    if (game) getCrawlStatisticHandler();
+  }, [game]);
 
   useEffect(() => {
     if (width > 692) {
@@ -126,10 +156,6 @@ export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncRes
             setStep(2);
           } else {
             setStep(3);
-            onSyncResult({
-              game: data.data.game ?? game,
-              scores: data.data.scores || [],
-            });
             if (data.data.status === "failed") {
               openAlertModal("同步游戏数据失败", "你的游戏数据同步时出现了错误，请查看同步结果了解详情。");
             } else if (data.data.status === "finished") {
@@ -148,105 +174,16 @@ export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncRes
     return () => {
       clearInterval(intervalId);
     };
-  }, [crawlStatus, idle, isLoggedOut, step, onSyncResult, game]);
-
-  const crawlStatusCard = !isLoggedOut ? (
-    <Card withBorder radius="md" className={classes.card} p="md" mt="xs">
-      <Card.Section className={classes.section}>
-        <Text size="xs" c="dimmed">
-          数据同步状态
-        </Text>
-        <Text
-          fz="lg"
-          c={crawlStatus ? (crawlStatus.status === "failed" ? "red" : (crawlStatus.status === "finished" ? "teal" : "default")) : "default"}
-        >
-          {!crawlStatus && "等待前置步骤完成"}
-          {crawlStatus && {
-            pending: "服务端正在爬取游戏数据",
-            finished: "游戏数据同步成功",
-            failed: "成绩同步不完全",
-          }[crawlStatus.status as "pending" | "finished" | "failed"]}
-        </Text>
-      </Card.Section>
-
-      {(!crawlStatus || (crawlStatus.status !== "finished" && crawlStatus.status !== "failed")) ? (
-        <Card.Section p="md">
-          <Text size="sm">
-            你的「{game === "maimai" ? "舞萌 DX" : "中二节奏"}」玩家信息与成绩将会被同步到 maimai DX 查分器，并与你的查分器账号绑定。
-          </Text>
-        </Card.Section>
-      ) : (
-        <>
-          <Card.Section className={classes.section}>
-            <Text fz="xs" c="dimmed">好友码</Text>
-            <Text fz="sm">{crawlStatus.friend_code}</Text>
-            <Group mt="md">
-              <div>
-                <Text fz="xs" c="dimmed">爬取耗时</Text>
-                <Text fz="sm">
-                  {Math.floor((new Date(crawlStatus.complete_time).getTime() - new Date(crawlStatus.create_time).getTime()) / 1000)} 秒
-                </Text>
-              </div>
-              <div>
-                <Group gap={4} align="center">
-                  <Text fz="xs" c="dimmed">爬取的成绩数</Text>
-                  <HoverCard width={280} shadow="md" withArrow>
-                    <HoverCard.Target>
-                      <ThemeIcon variant="subtle" color="gray" size="xs" style={{ cursor: "pointer" }}>
-                        <IconHelp />
-                      </ThemeIcon>
-                    </HoverCard.Target>
-                    <HoverCard.Dropdown>
-                      <Text size="sm">
-                        本次爬取过程中，服务器成功获取并<Mark>有变化</Mark>的成绩数量。
-                      </Text>
-                    </HoverCard.Dropdown>
-                  </HoverCard>
-                </Group>
-                <Text fz="sm">{(crawlStatus.scores || []).length}</Text>
-              </div>
-              <div>
-                <Text fz="xs" c="dimmed">爬取失败的难度</Text>
-                <Text fz="sm">
-                  {(!crawlStatus.failed_difficulties || crawlStatus.failed_difficulties.length === 0) ? "无" : crawlStatus.failed_difficulties.map((difficulty: number) => {
-                    return [
-                      "BASIC",
-                      "ADVANCED",
-                      "EXPERT",
-                      "MASTER",
-                      crawlStatus.game === "maimai" ? "Re:MASTER" : "ULTIMA",
-                    ][difficulty];
-                  }).join("、")}
-                </Text>
-              </div>
-            </Group>
-          </Card.Section>
-          <Card.Section p="md">
-            <SimpleGrid cols={small ? 2 : 4}>
-              <Button onClick={() => onViewSyncResult({ game: crawlStatus.game, scores: crawlStatus.scores || [] })}>
-                查看同步结果
-              </Button>
-              <Button variant="outline" onClick={() => navigate("/user/profile")}>
-                账号详情
-              </Button>
-              <Button variant="outline" onClick={() => navigate("/user/scores")}>
-                成绩管理
-              </Button>
-              <Button variant="outline" leftSection={<IconRepeat size={18} />} onClick={() => {
-                setCrawlStatus(null);
-                setStep(1);
-              }}>
-                重新同步
-              </Button>
-            </SimpleGrid>
-          </Card.Section>
-        </>
-      )}
-    </Card>
-  ) : null;
+  }, [crawlStatus, idle, isLoggedOut, step, game]);
 
   return (
     <>
+      <ScoresChangesModal
+        game={crawlStatus?.game ?? game}
+        scores={crawlStatus?.scores ?? []}
+        opened={resultModalOpened}
+        onClose={() => setResultModalOpened(false)}
+      />
       {(new Date()).getHours() >= 18 && (
         <Alert radius="md" icon={<IconAlertCircle />} title="游玩高峰期警告" color="yellow" mb="xl">
           <Text size="sm" mb="md">
@@ -359,7 +296,7 @@ export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncRes
                   { name: '中二节奏', description: '爬取玩家信息、成绩与收藏品', value: 'chunithm' },
                 ]}
                 value={game}
-                onChange={(value) => onGameChange(value as Game)}
+                onChange={(value) => setGame(value as typeof game)}
               />
               <Card className={classes.card} withBorder radius="md" mt="xs" mb="lg">
                 <Text fz="lg" fw={700}>服务器状态</Text>
@@ -385,7 +322,7 @@ export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncRes
           loading={isProxyAvailable && !crawlStatus}
           description={
             <Stack gap="xs" w={containerWidth}>
-              <Text fz="sm">复制微信 OAuth 链接，发送至安全的聊天中并打开</Text>
+                <Text fz="sm">复制微信 OAuth 链接，发送至安全的聊天中并打开</Text>
               {game && <WechatOAuthLink game={game} crawlToken={crawlToken} />}
               {!isLoggedOut && (
                 <Text>
@@ -396,13 +333,106 @@ export const ProxySyncSection = ({ crawlStatistic, game, onGameChange, onSyncRes
             </Stack>
           }
         />
-        <Stepper.Step
-          label="步骤 4"
-          loading={isProxyAvailable && crawlStatus?.status === "pending"}
-          description={<Text fz="sm">等待数据同步完成</Text>}
-        />
+      <Stepper.Step
+        label="步骤 4"
+        loading={isProxyAvailable && crawlStatus?.status === "pending"}
+        description={<Text fz="sm">等待数据同步完成</Text>}
+      />
       </Stepper>
-      {crawlStatusCard}
+      {!isLoggedOut && (
+        <Card withBorder radius="md" className={classes.card} p="md" mt="xs">
+          <Card.Section className={classes.section}>
+            <Text size="xs" c="dimmed">
+              数据同步状态
+            </Text>
+            <Text
+              fz="lg"
+              c={crawlStatus ? (crawlStatus.status === "failed" ? "red" : (crawlStatus.status === "finished" ? "teal" : "default")) : "default"}
+            >
+              {!crawlStatus && "等待前置步骤完成"}
+              {crawlStatus && {
+                pending: "服务端正在爬取游戏数据",
+                finished: "游戏数据同步成功",
+                failed: "成绩同步不完全",
+              }[crawlStatus.status as "pending" | "finished" | "failed"]}
+            </Text>
+          </Card.Section>
+
+          {(!crawlStatus || (crawlStatus.status !== "finished" && crawlStatus.status !== "failed")) ? (
+            <Card.Section p="md">
+              <Text size="sm">
+                你的「{game === "maimai" ? "舞萌 DX" : "中二节奏"}」玩家信息与成绩将会被同步到 maimai DX 查分器，并与你的查分器账号绑定。
+              </Text>
+            </Card.Section>
+          ) : (
+            <>
+              <Card.Section className={classes.section}>
+                <Text fz="xs" c="dimmed">好友码</Text>
+                <Text fz="sm">{crawlStatus.friend_code}</Text>
+                <Group mt="md">
+                  <div>
+                    <Text fz="xs" c="dimmed">爬取耗时</Text>
+                    <Text fz="sm">
+                      {Math.floor((new Date(crawlStatus.complete_time).getTime() - new Date(crawlStatus.create_time).getTime()) / 1000)} 秒
+                    </Text>
+                  </div>
+                  <div>
+                    <Group gap={4} align="center">
+                      <Text fz="xs" c="dimmed">爬取的成绩数</Text>
+                      <HoverCard width={280} shadow="md" withArrow>
+                        <HoverCard.Target>
+                          <ThemeIcon variant="subtle" color="gray" size="xs" style={{ cursor: "pointer" }}>
+                            <IconHelp />
+                          </ThemeIcon>
+                        </HoverCard.Target>
+                        <HoverCard.Dropdown>
+                          <Text size="sm">
+                            本次爬取过程中，服务器成功获取并<Mark>有变化</Mark>的成绩数量。
+                          </Text>
+                        </HoverCard.Dropdown>
+                      </HoverCard>
+                    </Group>
+                    <Text fz="sm">{(crawlStatus.scores || []).length}</Text>
+                  </div>
+                  <div>
+                    <Text fz="xs" c="dimmed">爬取失败的难度</Text>
+                    <Text fz="sm">
+                      {(!crawlStatus.failed_difficulties || crawlStatus.failed_difficulties.length === 0) ? "无" : crawlStatus.failed_difficulties.map((difficulty: number) => {
+                        return [
+                          "BASIC",
+                          "ADVANCED",
+                          "EXPERT",
+                          "MASTER",
+                          crawlStatus.game === "maimai" ? "Re:MASTER" : "ULTIMA",
+                        ][difficulty];
+                      }).join("、")}
+                    </Text>
+                  </div>
+                </Group>
+              </Card.Section>
+              <Card.Section p="md">
+                <SimpleGrid cols={small ? 2 : 4}>
+                  <Button onClick={() => setResultModalOpened(true)}>
+                    查看同步结果
+                  </Button>
+                  <Button variant="outline" onClick={() => navigate("/user/profile")}>
+                    账号详情
+                  </Button>
+                  <Button variant="outline" onClick={() => navigate("/user/scores")}>
+                    成绩管理
+                  </Button>
+                  <Button variant="outline" leftSection={<IconRepeat size={18} />} onClick={() => {
+                    setCrawlStatus(null);
+                    setStep(1);
+                  }}>
+                    重新同步
+                  </Button>
+                </SimpleGrid>
+              </Card.Section>
+            </>
+          )}
+        </Card>
+      )}
     </>
   );
 };
